+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [3.1.0](https://github.com/webbuilders-group/silverstripe-gridfield-calendar-view/tree/3.1.0) (2024-4-30)
+[Full Changelog](https://github.com/webbuilders-group/silverstripe-gridfield-calendar-view/compare/3.0.1...3.1.0)
+- `GridFieldCalendarView::handleCalendarFeed()` now returns a json string with the calendar data as under the `data` key, this a backwards compatible change
+- Changing the month/year in the calendar now properly updates the state in the url when the new response format is used
+
 ## [3.0.1](https://github.com/webbuilders-group/silverstripe-gridfield-calendar-view/tree/3.0.1) (2024-4-30)
 [Full Changelog](https://github.com/webbuilders-group/silverstripe-gridfield-calendar-view/compare/3.0.0...3.0.1)